pyproject.toml
[tool.poetry]
name = "lms-system"
version = "0.1.0"
description = ""
authors = ["KarelinIvan <ivan.karelin.1993@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
black = "^24.10.0"
django-filter = "^24.3"
filters = "^1.3.2"
djangorestframework-simplejwt = "^5.3.1"
celery = "^5.4.0"
coverage = "^7.6.8"
drf-yasg = "^1.21.8"
stripe = "^11.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
